fix for big sizes

+++ vsize.h
@@ -1,10 +1,11 @@
 #ifndef _PROC_VSIZE_H_
 #define _PROC_VSIZE_H_
 
+#include <stdint.h>
 #include <stdlib.h>
 #include <fcntl.h>
 #include <ctype.h>
 
-extern int getvsize ( int pid );
+extern size_t getvsize ( int pid );
 
 #endif // _PROC_VSIZE_H_
